// apps/frontend/src/types/GitSyncStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GitSyncState } from "./GitSyncState";

export type GitSyncStatus = { state: GitSyncState, remote: string | null, branch: string | null, localRevision: string | null, remoteRevision: string | null, ahead: number, behind: number, hasLocalChanges: boolean, };